fix(console): fix first query pod maybe get all
