+++ web/console/src/modules/cluster/components/resource/resourceDetail/ResourcePodTable.tsx
@@ -1,7 +1,9 @@
-import React, { useCallback, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import { Card, Pagination, Table, TableProps } from '@tea/component';
-import { PagingQuery } from '@tencent/ff-redux';
 import { RootProps } from '../../ClusterApp';
+import { router } from '../../../router';
+import { IsInNodeManageDetail } from './ResourceDetail';
+import { ResourceFilter } from '@src/modules/common';
 
 interface PodTableProps extends RootProps, TableProps {}
 
@@ -9,13 +11,40 @@ export function PodTabel({
   subRoot: {
     resourceDetailState: {
       podList,
-      podQuery: { paging, recordCount }
+      podQuery: { paging, recordCount },
+      resourceDetailInfo: { selection }
     }
   },
   actions,
   columns,
-  addons
+  addons,
+  route
 }: PodTableProps) {
+  useEffect(() => {
+    const urlParams = router.resolve(route);
+    const { type, resourceName } = urlParams;
+    const isInNodeManage = IsInNodeManageDetail(type);
+
+    if ((type === 'resource' || isInNodeManage) && resourceName !== 'cronjob') {
+      const { rid, clusterId } = route.queries;
+      let filter: ResourceFilter = {
+        regionId: +rid,
+        clusterId
+      };
+
+      if (!isInNodeManage) {
+        if (!selection) return;
+
+        filter = Object.assign(filter, {
+          namespace: route.queries['np'],
+          specificName: route.queries['resourceIns']
+        });
+      }
+      // 进行pod列表的轮询拉取
+      actions.resourceDetail.pod.poll(filter);
+    }
+  }, [selection, actions.resourceDetail.pod, route]);
+
   return (
     <Card>
       <Card.Body>
